refactor: extract S3 path functions into s3-path-utils.js
Move getS3Key, getS3KeyFromInfo, and getDocPathFromS3Key out of utils.js
into a dedicated s3-path-utils module. Add tests for all three functions.
